Improved support information gathering
