+++ go.mod
@@ -3,3 +3,8 @@ module github.com/essentialkaos/redis-latency-monitor
 go 1.18
 
 require github.com/essentialkaos/ek/v12 v12.111.1
+
+require (
+	github.com/essentialkaos/depsy v1.1.0 // indirect
+	golang.org/x/sys v0.18.0 // indirect
+)
